refactor(ui): split main screen
