+++ .github/PULL_REQUEST_TEMPLATE.md
@@ -8,6 +8,10 @@
 
 - [ ] `./gradlew quality`
 - [ ] Tested on device
+- [ ] Feature behavior has an appropriate test
+- [ ] Bug fix includes a regression test that failed before the fix
+- [ ] Regression test passes with the fix
+- [ ] Screenshot references were reviewed when the UI changed
 
 ## Related Issues
