+++ app/build.gradle
@@ -7,6 +7,7 @@ plugins {
     id 'org.jetbrains.kotlin.plugin.compose'
     id 'org.jetbrains.kotlin.plugin.serialization'
     id 'com.google.devtools.ksp'
+    id 'com.google.dagger.hilt.android'
     id 'com.github.triplet.play'
     id 'com.android.compose.screenshot'
 }
@@ -38,7 +39,7 @@ android {
         targetSdk 37
         versionCode 17
         versionName "0.7.0"
-        testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
+        testInstrumentationRunner "com.shapeshed.aerial.testing.AerialTestRunner"
     }
 
     signingConfigs {
@@ -211,6 +212,7 @@ dependencies {
     implementation 'androidx.activity:activity-compose:1.13.0'
     implementation 'androidx.lifecycle:lifecycle-viewmodel-compose:2.11.0'
     implementation 'androidx.lifecycle:lifecycle-runtime-compose:2.11.0'
+    implementation 'androidx.hilt:hilt-navigation-compose:1.4.0'
 
     def media3Version = '1.11.0'
     implementation "androidx.media3:media3-exoplayer:$media3Version"
@@ -233,6 +235,8 @@ dependencies {
     implementation 'com.squareup.okhttp3:okhttp:5.5.0'
     implementation 'androidx.datastore:datastore-preferences:1.2.1'
     implementation 'androidx.concurrent:concurrent-futures-ktx:1.3.0'
+    implementation 'com.google.dagger:hilt-android:2.60.1'
+    ksp 'com.google.dagger:hilt-android-compiler:2.60.1'
 
     testImplementation 'junit:junit:4.13.2'
     testImplementation 'org.json:json:20240303'
