+++ app/src/androidTest/java/com/shapeshed/aerial/ui/SearchHeaderTest.kt
@@ -12,12 +12,9 @@ import androidx.compose.ui.test.onAllNodesWithContentDescription
 import androidx.compose.ui.test.onNodeWithContentDescription
 import androidx.compose.ui.test.onNodeWithText
 import androidx.compose.ui.test.performClick
-import androidx.test.core.app.ApplicationProvider
-import com.shapeshed.aerial.AerialApp
 import com.shapeshed.aerial.MainActivity
 import com.shapeshed.aerial.R
-import kotlinx.coroutines.runBlocking
-import org.junit.Before
+import com.shapeshed.aerial.testing.AerialTestEnvironmentRule
 import org.junit.Rule
 import org.junit.Test
 
@@ -26,11 +23,8 @@ class SearchHeaderTest {
     @get:Rule
     val composeRule = createAndroidComposeRule<MainActivity>()
 
-    @Before
-    fun setUp() {
-        val app = ApplicationProvider.getApplicationContext<AerialApp>()
-        runBlocking { app.settingsDataStore.updateData { it.toMutablePreferences().apply { clear() } } }
-    }
+    @get:Rule
+    val testEnvironment = AerialTestEnvironmentRule()
 
     @Test
     fun rotatingMainScreenToLandscapeKeepsSearchCollapsed() {
