+++ app/src/main/java/com/shapeshed/aerial/MainActivity.kt
@@ -6,44 +6,24 @@ import androidx.activity.compose.setContent
 import androidx.activity.enableEdgeToEdge
 import androidx.activity.viewModels
 import androidx.core.splashscreen.SplashScreen.Companion.installSplashScreen
-import androidx.compose.runtime.remember
-import androidx.lifecycle.createSavedStateHandle
 import androidx.lifecycle.compose.dropUnlessResumed
-import androidx.lifecycle.viewmodel.compose.viewModel
-import androidx.lifecycle.viewmodel.initializer
-import androidx.lifecycle.viewmodel.viewModelFactory
+import androidx.hilt.lifecycle.viewmodel.compose.hiltViewModel
 import com.shapeshed.aerial.ui.MainScreen
 import com.shapeshed.aerial.ui.MainViewModel
 import com.shapeshed.aerial.ui.SettingsScreen
 import com.shapeshed.aerial.ui.SettingsViewModel
 import com.shapeshed.aerial.ui.StationEditScreen
 import com.shapeshed.aerial.ui.StationEditViewModel
-import com.shapeshed.aerial.ui.ZipSettingsBackupManager
+import com.shapeshed.aerial.ui.importStationLogo
 import com.shapeshed.aerial.ui.theme.AerialTheme
+import dagger.hilt.android.AndroidEntryPoint
 
+@AndroidEntryPoint
 class MainActivity : AppCompatActivity() {
 
-    private val mainViewModel: MainViewModel by viewModels {
-        val app = application as AerialApp
-        viewModelFactory {
-            initializer {
-                MainViewModel(app, app.repository, app.registryRepository, app.settingsDataStore, createSavedStateHandle())
-            }
-        }
-    }
+    private val mainViewModel: MainViewModel by viewModels()
 
-    private val settingsViewModel: SettingsViewModel by viewModels {
-        val app = application as AerialApp
-        viewModelFactory {
-            initializer {
-                SettingsViewModel(
-                    app,
-                    app.settingsDataStore,
-                    ZipSettingsBackupManager(app, app.repository, app.settingsDataStore),
-                )
-            }
-        }
-    }
+    private val settingsViewModel: SettingsViewModel by viewModels()
 
     override fun onCreate(savedInstanceState: Bundle?) {
         val splashScreen = installSplashScreen()
@@ -52,9 +32,6 @@ class MainActivity : AppCompatActivity() {
         enableEdgeToEdge()
         setContent {
             AerialTheme {
-                val repository = remember { (application as AerialApp).repository }
-                val registryRepository = remember { (application as AerialApp).registryRepository }
-
                 MainScreen(
                     viewModel = mainViewModel,
                     settingsContent = { onDismiss ->
@@ -64,16 +41,10 @@ class MainActivity : AppCompatActivity() {
                         )
                     },
                     stationEditContent = { stationId, onDismiss ->
-                        val vm: StationEditViewModel = viewModel(
-                            factory = viewModelFactory {
-                                initializer {
-                                    StationEditViewModel(
-                                        repository,
-                                        registryRepository,
-                                        stationId,
-                                    )
-                                }
-                            }
+                        val vm: StationEditViewModel = hiltViewModel(
+                            creationCallback = { factory: StationEditViewModel.Factory ->
+                                factory.create(stationId, ::importStationLogo)
+                            },
                         )
                         StationEditScreen(
                             viewModel = vm,
