+++ build.gradle
@@ -3,6 +3,7 @@ plugins {
     id 'org.jetbrains.kotlin.plugin.compose' version '2.4.10' apply false
     id 'org.jetbrains.kotlin.plugin.serialization' version '2.4.10' apply false
     id 'com.google.devtools.ksp' version '2.3.11' apply false
+    id 'com.google.dagger.hilt.android' version '2.60.1' apply false
     id 'com.github.triplet.play' version '4.1.1' apply false
     id 'com.android.compose.screenshot' version '0.0.1-alpha16' apply false
 }
